Restrict overflow of page content on slide-in/slide-out
Fixed an issue where the page width would expand with the slide out animation of the page content. It was unnoticable on desktop browsers but in mobile browsers the page would zoom out quickly to fit in all of the page content.

Added jquery.easing. Not sure if correctly implemented in actions.js